% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.AlkAnilineSeq.R
\docType{data}
\name{RNAmodR.AlkAnilineSeq-datasets}
\alias{RNAmodR.AlkAnilineSeq-datasets}
\alias{msaas}
\title{Example data in the RNAmodR.AlkAnilineSeq package}
\format{a \code{ModSetAlkAnilineSeq} instance}
\usage{
data(msaas)
}
\description{
This contains an example ModifierSet object of type ModSetAlkAnilineSeq
}
\keyword{datasets}
