% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.AlkAnilineSeq.R
\docType{package}
\name{RNAmodR.AlkAnilineSeq-package}
\alias{RNAmodR.AlkAnilineSeq-package}
\title{RNAmodR.AlkAnilineSeq: Detection of m7G, m3C and D modification by AlkAnilineSeq}
\description{
RNAmodR.AlkAnilineSeq implements the detection of m7G, m3C and D modifications on RNA from experimental data generated with the AlkAnilineSeq protocol. The package builds on the core functionality of the RNAmodR package to detect specific patterns of the modifications in high throughput sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FelixErnst/RNAmodR.AlkAnilineSeq}
  \item Report bugs at \url{https://github.com/FelixErnst/RNAmodR.AlkAnilineSeq/issues}
}

}
\author{
\strong{Maintainer}: Felix G.M. Ernst \email{felix.gm.ernst@outlook.com} (\href{https://orcid.org/0000-0001-5064-0928}{ORCID})

Other contributors:
\itemize{
  \item Denis L.J. Lafontaine \email{denis.lafontaine@ulb.ac.be} [contributor, funder]
}

}
\keyword{internal}
