% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.AlkAnilineSeq.R
\name{RNAmodR.AlkAnilineSeq}
\alias{RNAmodR.AlkAnilineSeq}
\title{RNAmodR.AlkAnilineSeq}
\description{
`RNAmodR.AlkAnilineSeq` implements the detection of 7-methyl guanosine,
3-methyl cytidine and dihydrouridine from AlkAnilineseq data using the 
workflow and class the package `RNAmodR` provides.
}
\seealso{
Further details are described in the man pages of the 
\code{\link[RNAmodR:Modifier-class]{Modifier}} object and the vignettes.
}
\author{
Felix G M Ernst [aut], Denis L J Lafontaine [fnd]
}
