% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-AlkAnilineSeq-class.R,
%   R/Modifier-AlkAnilineSeq-viz.R
\docType{methods}
\name{ModAlkAnilineSeq-functions}
\alias{ModAlkAnilineSeq-functions}
\alias{aggregate}
\alias{modify}
\alias{settings}
\alias{visualizeData}
\alias{visualizeDataByCoord}
\alias{settings<-,ModAlkAnilineSeq-method}
\alias{aggregateData,ModAlkAnilineSeq-method}
\alias{findMod,ModAlkAnilineSeq-method}
\alias{getDataTrack,ModAlkAnilineSeq-method}
\alias{visualizeDataByCoord,ModAlkAnilineSeq,GRanges-method}
\alias{visualizeData,ModAlkAnilineSeq-method}
\alias{visualizeDataByCoord,ModSetAlkAnilineSeq,GRanges-method}
\alias{visualizeData,ModSetAlkAnilineSeq-method}
\title{Functions for ModAlkAnilineSeq}
\usage{
\S4method{settings}{ModAlkAnilineSeq}(x) <- value

\S4method{aggregateData}{ModAlkAnilineSeq}(x)

\S4method{findMod}{ModAlkAnilineSeq}(x)

\S4method{getDataTrack}{ModAlkAnilineSeq}(x, name, type, ...)

\S4method{visualizeDataByCoord}{ModAlkAnilineSeq,GRanges}(x, coord,
  type = c("ends", "scoreNC", "scoreSR"), window.size = 15L, ...)

\S4method{visualizeData}{ModAlkAnilineSeq}(x, name, from = 1L,
  to = 30L, type = c("ends", "scoreNC", "scoreSR"), ...)

\S4method{visualizeDataByCoord}{ModSetAlkAnilineSeq,GRanges}(x, coord,
  type = c("scoreNC", "scoreSR", "ends"), window.size = 15L, ...)

\S4method{visualizeData}{ModSetAlkAnilineSeq}(x, name, from = 1L,
  to = 30L, type = c("scoreNC", "scoreSR", "ends"), ...)
}
\arguments{
\item{x}{a \code{\link[RNAmodR:Modifier-class]{Modifier}} or a
\code{\link[RNAmodR:ModifierSet-class]{ModifierSet}} object. For more
details see also the man pages for the functions mentioned below.}

\item{value}{See \code{\link[RNAmodR:Modifier-class]{settings}}}

\item{coord, name, from, to, type, window.size, ...}{See 
\code{\link[RNAmodR:visualizeData]{visualizeData}}.}

\item{force}{See \code{\link[RNAmodR:aggregate]{aggregate}}}
}
\description{
All of the functions of \code{\link[RNAmodR:Modifier-class]{Modifier}} and
the \code{\link[RNAmodR:ModifierSet-class]{ModifierSet}} classes are
inherited by the \code{ModAlkAnilineSeq} and \code{ModSetAlkAnilineSeq}
classes.
}
\details{
\code{ModAlkAnilineSeq} specific arguments for \link{visualizeData}:
\itemize{
\item{\code{colour} - }{a named character vector of \code{length = 4} 
for the colours of the individual histograms. The names are expected to be 
\code{c("scoreNC","scoreSR")}}
}
}
