% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-AlkAnilineSeq-class.R,
%   R/Modifier-AlkAnilineSeq-viz.R
\docType{class}
\name{ModAlkAnilineSeq}
\alias{ModAlkAnilineSeq}
\alias{AlkAnilineSeq}
\alias{ModAlkAnilineSeq-class}
\alias{ModAlkAnilineSeq,character-method}
\alias{ModAlkAnilineSeq,BamFileList-method}
\alias{ModAlkAnilineSeq,SequenceData-method}
\alias{modify,ModAlkAnilineSeq-method}
\alias{ModSetAlkAnilineSeq-class}
\alias{ModSetAlkAnilineSeq}
\alias{visualizeData}
\alias{visualizeDataByCoord,ModAlkAnilineSeq,GRanges-method}
\alias{visualizeData,ModAlkAnilineSeq-method}
\alias{visualizeDataByCoord,ModSetAlkAnilineSeq,GRanges-method}
\alias{visualizeData,ModSetAlkAnilineSeq-method}
\title{ModAlkAnilineSeq}
\usage{
\S4method{settings}{ModAlkAnilineSeq}(x) <- value

\S4method{ModAlkAnilineSeq}{character}(x, fasta, gff,
  modifications = NULL, ...)

\S4method{ModAlkAnilineSeq}{BamFileList}(x, fasta, gff,
  modifications = NULL, ...)

\S4method{ModAlkAnilineSeq}{SequenceData}(x, modifications = NULL, ...)

\S4method{aggregate}{ModAlkAnilineSeq}(x, force = FALSE)

\S4method{modify}{ModAlkAnilineSeq}(x, force = FALSE)

ModSetAlkAnilineSeq(x, fasta = NA, gff = NA)

\S4method{visualizeDataByCoord}{ModAlkAnilineSeq,GRanges}(x, coord,
  type = c("ends", "scoreNC", "scoreSR"), window.size = 15L, ...)

\S4method{visualizeData}{ModAlkAnilineSeq}(x, name, from = 1L,
  to = 30L, type = c("ends", "scoreNC", "scoreSR"), ...)

\S4method{visualizeDataByCoord}{ModSetAlkAnilineSeq,GRanges}(x, coord,
  type = c("scoreNC", "scoreSR", "ends"), window.size = 15L, ...)

\S4method{visualizeData}{ModSetAlkAnilineSeq}(x, name, from = 1L,
  to = 30L, type = c("scoreNC", "scoreSR", "ends"), ...)
}
\arguments{
\item{...}{Optional arguments overwriting default values, which are
\itemize{
\item{minSignal:}{The minimal singal at the position as integer value 
(default: \code{minSignal = 10L}). If the reaction is very specific a lower
value may need to be used}
\item{minScore:}{minimum for score to identify m7G, m3C and D positions 
de novo (default: \code{minScore = 0.5})}
\item{maxLength:}{The default read length. Reads with this length or longer
are discarded, since they represent non-fragemented reads. This might need to
be adjusted for individual samples dending on the experimental conditions.
This is argument is passed on to \code{\link{ProtectedEndSequenceData}} 
(default: \code{maxLength = 50L})}
\item{other arguments}{which are passed on to 
\code{\link{End5SequenceData}}}
}}
}
\description{
title
}
\details{
\code{ModAlkAnilineSeq} specific arguments for \link{visualizeData}:
\itemize{

}
}
\author{
Felix G.M. Ernst \email{felix.gm.ernst@outlook.com}
}
