% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.AlkAnilineSeq.R
\docType{package}
\name{RNAmodR.AlkAnilineSeq}
\alias{RNAmodR.AlkAnilineSeq}
\alias{RNAmodR.AlkAnilineSeq-package}
\title{RNAmodR.AlkAnilineSeq}
\description{
todo xczCZXcZZcc cCX
}
\author{
Felix G M Ernst [aut]
}
